Implemented compare with your result observer strategy

+++ lib/model/activity/record_activity_widget_model.dart
@@ -1,5 +1,5 @@
 class RecordActivityWidgetModel {
-  final int currentDistanceInKm;
+  final double currentDistanceInKm;
   final int currentPlayerPosition;
   final List<RecordActivityWidgetRankingItem> ranking;
